Add files via upload
In this assignment, the program is python 3.7 version.

1: Initialization
In this part, I create class for each peer in main function. Every time a peer created, the functions for UDP server, TCP server and UDP client from the class will run in multithreading. The UDP server and TCP server will always running, and the keyboard detect also work as another thread in each peer. In TCP server, it will check three messages that are request file, peer quit or peer killed. In UDP server, it will check message is ping or file transmission. All the servers’ recv is 1024.
2: Ping successors
After the UDP client(PingClient) built, it will always send ping message to 2 successors. Since UDP server(PingServer) get the message of ping, it responses the predecessors. At the same time, it will record the predecessor which will be used when in step 4 peer departure. 
3: Requesting file
(1)	Find file position
In this part, TCPClient function always used to request successor to find responding peer, when the responding peer is found, it will run another TCP client function named tcp_respondingpeer to connect requesting peer.
When the check_input function detect the keyboard input is request filename, it will run the TCPClient function which will send message to request file to successor over TCP, until finding the responding peer. All the message formats are string containing file number, request peer id. Every time the peer will check file’s hash value with hash function. When the responding peer find it has the file, it will run another TCP client function tcp_respondingpeer to send a message to the requesting peer. The requesting peer will check the message receive in TCP server, and then response. And there is a global flag in TCP server for writing file set 1. So, the UDP server in requesting peer will know there will be file transmission and run the file writing code.
(2)	File transmission
When the responding peer get the response from requesting peer, it runs another UPD client function named udp_source in the code with threading for sending file. In this function, it will open the file and separate it into several packets depending on MSS. The timeout is 1s. Each packet contains a header and file message. The header is 40bytes with seq number and ack number and start time. The file message is as big as the MSS.
Every time when the responding peer get a response of packet acknowledged from requesting peer, the seq number and the ack number will add 1 in header and the flag of packets number will reduce 1 and until 0 the loop of sending files will stop as well as updating packet. If time out, the flag and the packet are not change and retransmission happen. The random rate number is also gotten at the beginning of the loop and comparing with drop rate. In the requesting peer, it is similar except recording drop and rtx.

4: Peer departure
When input the quit, the check_input function will set a global flag for depart, then the TCP_depart function will run in UDP server by using predecessor id. So the predecessors will receive message of depart peer id and the successors id of it. Therefore, the peer will send a response to the depart peer and change successor. The depart peer will stop send ping after getting response.

5: Kill a peer
Setting 10s time out for receive response and 5s sleeping time for sending ping in PingClient function. There are two flag for the first predecessor(nearest one) and the second predecessor(the far one) of the killed peer. Flag successor1_kill and successor2_kill add 1 every time time out happen until they equal 3. Function TCP_kill_2 and function TCP_kill_1 will run and send message to their nearest successor. After function TCP_kill_1 run, the peer connect with its second successor and then update successors. When the successors updated, the global flag kill_1_change set 1 and a response will send to the second predecessor when kill_1_change =1 . For the second predecessor, when it get the response from the first one, it updates successor. The new successor information will send as string.
For the three parameters:
the number of consecutive ping request messages that a peer fails to respond: 3 times
the frequency of the ping messages: 5 s
timeout: 10 s.


In PingClient function: 
When a peer killed and time out happen in predecessors, flag successor1_kill and successor2_kill will be set separately in two ping sending response code.